+++ app/src/main/java/com/forcetower/uefs/feature/setup/IntroductionFragment.kt
@@ -153,7 +153,7 @@ class IntroductionFragment: UFragment(), Injectable {
                             .setBorderLineColor(ac)
                             .setBorderCornerColor(ac)
                             .setActivityMenuIconColor(ac)
-                            .setBorderLineThickness(getPixelsFromDp(requireContext(), 2).toFloat())
+                            .setBorderLineThickness(getPixelsFromDp(requireContext(), 2))
                             .setActivityTitle(getString(R.string.cut_profile_image))
                             .setGuidelines(CropImageView.Guidelines.OFF)
                             .start(requireContext(), this)
